references fix, and move to env.
