+++ python-backend/.env-example
@@ -22,3 +22,6 @@ export ELASTIC_ENABLED=0
 export ELASTIC_SERVICE_NAME=Local-Dev
 export ELASTIC_SECRET_TOKEN="abcdefg"
 export ELASTIC_SERVER_URL="https://localhost:8200"
+
+export LDAP_IDIR_USERNAME=fnlastn
+export LDAP_IDIR_PASSWORD=idirpw
