+++ python-backend/app/scheduled_jobs/NRIS_jobs.py
@@ -1,6 +1,6 @@
 import requests
 from app.extensions import cache, sched
-from app.api.services.nris_services import NRIS_service
+from app.api.services import NRIS_service
 from app.api.mines.mine.models.mine import Mine
 from app.api.constants import NRIS_JOB_PREFIX, NRIS_MMLIST_JOB, NRIS_MAJOR_MINE_LIST, TIMEOUT_24_HOURS, TIMEOUT_60_MINUTES
 from app.api.utils.apm import register_apm
